% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sox-package.R
\docType{package}
\name{sox-package}
\alias{sox-package}
\title{sox}
\description{
Efficient procedures for fitting and cross-validating the structurally-regularized time-dependent Cox models. The penalty term is a weighted sum of infinity norms of (overlapping) groups of coefficients, which can select variables following a user-specified grouping structure.
}
