% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sox-package.R
\docType{package}
\name{sox-package}
\alias{sox-package}
\alias{_PACKAGE}
\title{sox: Structured Learning in Time-Dependent Cox Models}
\description{
Efficient procedures for fitting and cross-validating the structurally-regularized time-dependent Cox models.
}
\author{
\strong{Maintainer}: Yi Lian \email{yi.lian@mail.mcgill.ca}

Authors:
\itemize{
  \item Guanbo Wang
  \item Archer Y. Yang
  \item Mireille E. Schnitzer
  \item Robert W. Platt
  \item Rui Wang
  \item Marc Dorais
  \item Sylvie Perreault
}

Other contributors:
\itemize{
  \item Julien Mairal [contributor]
  \item Yuansi Chen [contributor]
}

}
\keyword{internal}
